Don't duplicate code for small-error
