+++ failure-1.X/src/lib.rs
@@ -56,7 +56,6 @@ with_std! {
     mod sync_failure;
     pub use sync_failure::SyncFailure;
 
-    #[cfg_attr(feature = "small-error", path = "./small_error.rs")]
     mod error;
 
     use std::error::Error as StdError;
